requirements.txt
absl-py==1.0.0
aiohttp==3.8.4
aiosignal==1.2.0
altair==4.2.0
amqp==5.1.1
anyio==3.6.2
argcomplete==1.10.0
asgiref==3.6.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
autocorrect==0.4.4
Automat==22.10.0
autopep8==1.5.7
base58==2.1.1
bcrypt==4.0.1
beautifulsoup4==4.8.2
billiard==3.6.4.0
bitarray==1.2.2
black==21.11b1
blinker==1.5
blis==0.7.9
Bottleneck==1.3.2
bpemb==0.3.4
bs4==0.0.1
CacheControl==0.12.11
cachetools==4.2.4
catalogue==1.0.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.9
chatgptpy==1.0.8
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
commonmark==0.9.1
compressed-rtf==1.0.6
confection==0.0.4
conllu==4.5.2
constantly==15.1.0
contractions==0.0.21
coverage==7.2.3
crispy-tailwind==0.5.0
cryptography==39.0.2
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.10.0
cymem==2.0.7
cytoolz==0.11.2
dataclasses-json==0.5.7
dateparser==1.1.0
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
Django==2.2
django-bootstrap4==23.1
django-cors-headers==4.0.0
django-crispy-forms==1.7.2
djangorestframework==3.14.0
dnspython==2.3.0
docker==6.0.1
docopt==0.6.2
docx2txt==0.8
duckduckgo-search==2.8.6
ebcdic==1.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.4
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
exceptiongroup==1.1.1
extract-msg==0.28.7
fastai==1.0.61
fastapi==0.89.1
fastprogress==1.0.2
filelock==3.7.1
firebase-admin==6.1.0
flair==0.11.3
flake8==6.0.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
flatbuffers==2.0
frozenlist==1.2.0
fst-pso==1.8.1
ftfy==6.1.1
future==0.18.3
FuzzyTM==2.0.5
gast==0.4.0
gcloud==0.18.3
gdown==4.4.0
gensim==3.8.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
GoogleBard==1.0.0
greenlet==2.0.1
grpcio==1.51.3
grpcio-status==1.51.3
gTTS==2.3.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.6.0
hexbytes==0.2.2
httplib2==0.21.0
huggingface-hub==0.12.0
hyperlink==21.0.0
hyperopt==0.2.7
identify==2.5.22
idna==3.3
imageio==2.16.0
IMAPClient==2.1.0
importlib-metadata==3.10.1
incremental==22.10.0
inflect==6.0.2
iniconfig==1.1.1
install==1.3.5
ipfshttpclient==0.8.0a2
isort==5.12.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Janome==0.4.2
jellyfish==0.9.0
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.0.1
Js2Py==0.74
json5==0.9.11
jsonschema==3.2.0
jwcrypto==1.4.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.2.4
konoha==4.6.5
langchain==0.0.174
langcodes==3.3.0
langdetect==1.0.9
libclang==13.0.0
linkedin-scraper==2.11.0
loguru==0.7.0
lru-dict==1.1.7
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.4.3
mccabe==0.7.0
miniful==0.0.6
more-itertools==9.0.0
mpld3==0.3
msgpack==1.0.5
multiaddr==0.0.9
multidict==5.2.0
multitasking==0.0.10
murmurhash==1.0.9
mypy==0.991
mypy-extensions==0.4.3
netaddr==0.8.0
networkx==2.6.3
nltk==3.8.1
nodeenv==1.7.0
numexpr==2.8.4
numpy==1.21.3
nvidia-ml-py3==7.352.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
openai==0.27.2
openapi-schema-pydantic==1.2.4
openpyxl==3.0.9
opt-einsum==3.3.0
orjson==3.8.10
outcome==1.2.0
overrides==3.1.0
packaging==21.3
pafy==0.5.5
pandas==1.3.4
parsel==1.8.1
parsimonious==0.8.1
pathspec==0.9.0
pathy==0.10.1
patsy==0.5.3
pdfkit==1.0.0
pdfminer.six==20191110
pdfminer3==2018.12.3.0
pi==0.1.2
Pillow==9.0.1
pinecone-client==2.2.1
pipenv==2022.12.19
pipwin==0.5.2
pipx==0.16.4
plac==1.1.3
platformdirs==2.4.0
playsound==1.2.2
plotly==5.12.0
plotly-express==0.4.1
pluggy==1.0.0
pptree==3.1
pre-commit==3.2.2
preshed==3.0.8
prompt-toolkit==3.0.38
Protego==0.2.1
proto-plus==1.22.2
protobuf==4.22.1
py==1.11.0
py-solc-x==1.1.1
py4j==0.10.9.7
pyahocorasick==2.0.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.8.2
pydeck==0.8.0
PyDispatcher==2.0.7
pyflakes==3.0.1
pyFUME==0.2.25
Pygments==2.14.0
pyjsparser==2.7.1
PyJWT==2.6.0
Pympler==1.0.1
PyMySQL==1.0.2
pynvx==1.0.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
PyPDF2==1.26.0
PyPrind==2.11.3
Pyrebase4==4.6.0
pyresparser==1.0.6
pyrsistent==0.18.0
pySmartDL==1.3.4
PySocks==1.7.1
pytest==6.2.5
pytest-mock==3.10.0
python-binance==1.0.15
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==1.0.0
python-jwt==4.0.0
python-pptx==0.6.18
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0
queuelib==1.6.2
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
reportlab==3.6.12
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
rich==13.1.0
rlp==2.0.1
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.1
scipy==1.7.0
scrape==0.11.3
scrapeops-scrapy==0.5.2
scrapeops-scrapy-proxy-sdk==1.0
Scrapy==2.8.0
seaborn==0.11.1
segtok==1.5.11
selenium==4.9.0
semantic-version==2.8.5
semver==2.13.0
sentencepiece==0.1.95
service-identity==21.1.0
simpful==2.9.0
six==1.12.0
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sourcery==1.2.0
spacy==2.3.5
spacy-legacy==3.0.11
spacy-loggers==1.0.4
SpeechRecognition==3.8.1
SQLAlchemy==1.4.46
sqlitedict==2.1.0
sqlparse==0.4.3
srsly==1.0.6
starlette==0.22.0
statsmodels==0.13.5
streamlit==1.17.0
streamlit-tags==1.2.8
stripe==5.0.0
svglib==1.5.1
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
textract==1.6.5
textsearch==0.0.17
thinc==7.4.6
threadpoolctl==2.2.0
tifffile==2022.2.9
tika==2.6.0
tiktoken==0.3.3
tinycss2==1.2.1
tld==0.13
tldextract==3.3.1
tls-client==0.1.9
tokenizers==0.13.2
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch==1.10.0
torch-tb-profiler==0.3.1
torchvision==0.11.1
tornado==6.2
tqdm==4.65.0
transformers==4.27.4
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.10.0
Twisted==22.10.0
twitter-application-only-auth==0.3.3
typer==0.4.2
typing-inspect==0.8.0
typing_extensions==4.4.0
tzdata==2021.5
tzlocal==4.2
ujson==5.1.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.15
userpath==1.7.0
uvicorn==0.20.0
validators==0.20.0
varint==1.0.2
vine==5.0.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
w3lib==2.1.1
wasabi==0.10.1
watchdog==2.2.1
wcwidth==0.2.6
web3==5.25.0
webdriver-manager==3.8.6
webencodings==0.5.1
webparser==1.0
websocket-client==1.5.1
websockets==9.1
Werkzeug==2.2.2
wfdb==3.4.1
Wikipedia-API==0.5.8
wrapt==1.13.3
wsproto==1.2.0
WTForms==3.0.1
xlrd==1.2.0
XlsxWriter==3.0.7
yake==0.4.8
yarl==1.7.2
yfinance==0.1.70
youtube-dl==2021.12.17
zipp==3.7.0
zope.interface==6.0
